Add Service , AppService for Admin,Order,Service,CategoryService,ServiceSubCategory

+++ HomeService.Domain.Core/DTOs/CommentCreateDto.cs
@@ -9,7 +9,7 @@ public class CommentCreateDto
     [DisplayName("متن کامنت")]
     public string Description { get; set; }
     [DisplayName("رضایت")]
-    public float Score { get; set; }
+    public int Score { get; set; }
     public int CustomerId { get; set; }
     public Customer Customer { get; set; }
     public int ExpertId { get; set; }
